Move all the handlers into the use app hook. So now we need to wire them all back up into the app page.
